Refactor: Split CreateBtn from FamPage
